# [Backport to 5.12] Pre-compiled regexp optimization